refactor(components): move header related components into dedicated directory
